# nursery-inventory-system/nursery-inventory-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plants:
    get:
      operationId: listPlants
      tags: [plants]
      summary: List all plants
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: plantType
          schema: { type: string }
        - in: query
          name: potSize
          schema: { type: string }
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"

    post:
      operationId: createPlant
      tags: [plants]
      summary: Create a plant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /plants/{id}:
    get:
      operationId: getPlant
      tags: [plants]
      summary: Get a plant
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Plant record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    put:
      operationId: updatePlant
      tags: [plants]
      summary: Update a plant
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deletePlant
      tags: [plants]
      summary: Delete a plant
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Plant Batches ─────────────────────────────────────────────────────────
...
  /production-batches:
    get:
      operationId: listProductionBatches
      tags: [production-batches]
      summary: List all production batches
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Search across productionBatchNumber, productionType, plant name
        - in: query
          name: plantId
          schema: { type: integer }
        - in: query
          name: nurseryLocation
          schema: { type: string }
        - in: query
          name: productionType
          schema: { type: string }
      responses:
        "200":
          description: List of production batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionBatch"

    post:
      operationId: createProductionBatch
      tags: [production-batches]
      summary: Create a production batch (grafting run)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionBatchInput"
      responses:
        "201":
          description: Created production batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Production batch number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /production-batches/{id}:
    get:
      operationId: getProductionBatch
      tags: [production-batches]
      summary: Get a production batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Production batch record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    put:
      operationId: updateProductionBatch
      tags: [production-batches]
      summary: Update a production batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionBatchUpdate"
      responses:
        "200":
          description: Updated production batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Production batch number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteProductionBatch
      tags: [production-batches]
      summary: Delete a production batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Projects ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Plant Master schemas ─────────────────────────────────────────────────
    Plant:
      type: object
      required: [id, plantCode, botanicalName, commonName, category, plantType, potSize, createdAt, updatedAt]
      properties:
        id:           { type: integer }
        plantCode:    { type: string }
        botanicalName: { type: string }
        commonName:   { type: string }
        category:     { type: string }
        plantType:    { type: string }
        potSize:      { type: string }
        imageUrl:     { type: ["string", "null"] }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
    PlantInput:
      type: object
      required: [plantCode, botanicalName, commonName, category, plantType, potSize]
      properties:
        plantCode:    { type: string, minLength: 1 }
        botanicalName: { type: string, minLength: 1 }
        commonName:   { type: string, minLength: 1 }
        category:     { type: string, minLength: 1 }
        plantType:    { type: string, minLength: 1 }
        potSize:      { type: string, minLength: 1 }
        imageUrl:     { type: ["string", "null"] }
    PlantUpdate:
      type: object
      properties:
        plantCode:    { type: string, minLength: 1 }
        botanicalName: { type: string, minLength: 1 }
        commonName:   { type: string, minLength: 1 }
        category:     { type: string, minLength: 1 }
        plantType:    { type: string, minLength: 1 }
        potSize:      { type: string, minLength: 1 }
        imageUrl:     { type: ["string", "null"] }

    # ── Plant Batch schemas ──────────────────────────────────────────────────
...
    ProductionBatchPlant:
      type: object
      required: [id, plantCode, commonName, botanicalName, potSize]
      properties:
        id:            { type: integer }
        plantCode:     { type: string }
        commonName:    { type: string }
        botanicalName: { type: string }
        potSize:       { type: string }
        imageUrl:      { type: ["string", "null"] }
    ProductionBatch:
      type: object
      required:
        - id
        - productionBatchNumber
        - plantId
        - plant
        - productionType
        - rootstockQuantity
        - successfulPlants
        - failedPlants
        - laborCost
        - potCost
        - soilCost
        - fertilizerCost
        - chemicalCost
        - waterCost
        - otherCosts
        - totalProductionCost
        - costPerPlant
        - nurseryLocation
        - productionDate
        - createdAt
        - updatedAt
      properties:
        id:                    { type: integer }
        productionBatchNumber: { type: string }
        plantId:               { type: integer }
        plant:
          $ref: "#/components/schemas/ProductionBatchPlant"
        productionType:        { type: string }
        rootstockQuantity:     { type: integer }
        successfulPlants:      { type: integer }
        failedPlants:          { type: integer }
        laborCost:             { type: number }
        potCost:               { type: number }
        soilCost:              { type: number }
        fertilizerCost:        { type: number }
        chemicalCost:          { type: number }
        waterCost:             { type: number }
        otherCosts:            { type: number }
        totalProductionCost:   { type: number }
        costPerPlant:          { type: number }
        nurseryLocation:       { type: string }
        productionDate:        { type: string, format: date }
        createdAt:             { type: string, format: date-time }
        updatedAt:             { type: string, format: date-time }
    ProductionBatchInput:
      type: object
      required:
        - plantId
        - productionType
        - rootstockQuantity
        - successfulPlants
        - productionDate
        - nurseryLocation
      properties:
        productionBatchNumber: { type: string }
        plantId:               { type: integer }
        productionType:        { type: string, minLength: 1 }
        rootstockQuantity:     { type: integer, minimum: 1 }
        successfulPlants:      { type: integer, minimum: 0 }
        failedPlants:          { type: integer, minimum: 0 }
        laborCost:             { type: number, minimum: 0 }
        potCost:               { type: number, minimum: 0 }
        soilCost:              { type: number, minimum: 0 }
        fertilizerCost:        { type: number, minimum: 0 }
        chemicalCost:          { type: number, minimum: 0 }
        waterCost:             { type: number, minimum: 0 }
        otherCosts:            { type: number, minimum: 0 }
        productionDate:        { type: string, format: date }
        nurseryLocation:       { type: string, minLength: 1 }
    ProductionBatchUpdate:
      type: object
      properties:
        productionBatchNumber: { type: string, minLength: 1 }
        plantId:               { type: integer }
        productionType:        { type: string, minLength: 1 }
        rootstockQuantity:     { type: integer, minimum: 1 }
        successfulPlants:      { type: integer, minimum: 0 }
        failedPlants:          { type: integer, minimum: 0 }
        laborCost:             { type: number, minimum: 0 }
        potCost:               { type: number, minimum: 0 }
        soilCost:              { type: number, minimum: 0 }
        fertilizerCost:        { type: number, minimum: 0 }
        chemicalCost:          { type: number, minimum: 0 }
        waterCost:             { type: number, minimum: 0 }
        otherCosts:            { type: number, minimum: 0 }
        productionDate:        { type: string, format: date }
        nurseryLocation:       { type: string, minLength: 1 }

    # ── Project schemas ──────────────────────────────────────────────────────
...